Scope router cache lifetime fix to TrainerRank model instances
