+++ src/art/trainer_rank/_impl.py
@@ -23,7 +23,7 @@ from pathlib import Path
 import struct
 import threading
 import time
-from types import TracebackType
+from types import MethodType, TracebackType
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -1135,6 +1135,46 @@ def _split_chunks(
     return tuple(tuple(chunk) for chunk in chunks)
 
 
+def _configure_moe_dispatcher_caches(model: Sequence[torch.nn.Module]) -> None:
+    for chunk in model:
+        for module in chunk.modules():
+            dispatcher = getattr(module, "token_dispatcher", None)
+            if dispatcher is None:
+                continue
+            from megatron.core.transformer.moe.token_dispatcher import (
+                MoEAlltoAllTokenDispatcher,
+            )
+
+            if type(dispatcher) is not MoEAlltoAllTokenDispatcher or (
+                "dispatch_preprocess" in vars(dispatcher)
+            ):
+                continue
+
+            def _dispatch_preprocess(
+                self: Any,
+                hidden_states: torch.Tensor,
+                routing_map: torch.Tensor,
+                probs: torch.Tensor,
+            ):
+                result = MoEAlltoAllTokenDispatcher.dispatch_preprocess(
+                    self, hidden_states, routing_map, probs
+                )
+                # MCore uses this cache only for dtype. Keep the real routing
+                # outputs differentiable without retaining their checkpoint graph.
+                self.probs = probs.new_empty(0)
+                return result
+
+            # Persist through caller-owned backward/checkpoint recomputation;
+            # other dispatcher instances and custom implementations stay intact.
+            setattr(
+                dispatcher,
+                "dispatch_preprocess",
+                MethodType(_dispatch_preprocess, dispatcher),
+            )
+            if (probs := getattr(dispatcher, "probs", None)) is not None:
+                dispatcher.probs = probs.new_empty(0)
+
+
 class TrainerRank:
     def __init__(self, runtime: TrainingRuntime) -> None:
         pp_size = int(getattr(runtime.provider, "pipeline_model_parallel_size", 1) or 1)
@@ -1271,6 +1311,7 @@ class TrainerRank:
         self._planning_seconds_accum = 0.0
         self._speculative_planning_seconds = 0.0
         self._last_forward_telemetry_snapshot: dict[str, Any] | None = None
+        _configure_moe_dispatcher_caches(runtime.model)
         self.zero_grad()
 
     def zero_grad(self) -> None:
